+++ src/common/network-switcher/index.ts
@@ -1,0 +1,3 @@
+export { default as BookshelfNetworkSwitcher } from '@/common/network-switcher/BookshelfNetworkSwitcher.vue'
+export { default as HeaderNetworkSwitcher } from '@/common/network-switcher/HeaderNetworkSwitcher.vue'
+export { default as NetworkItem } from '@/common/network-switcher/NetworkItem.vue'