+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 #### Added
-- Network switcher component(doing nothing for now)
+- Network switcher component in bookshelft(doing nothing for now)
+- Network switcher component in header
+- User drop-down menu
 
 #### Changed
 - About page redesign